Refactor S8 GEMM/V JIT kernel handling for col-major operations (#346)
- Introduced temporary variables for input transformation in JIT kernel
  generation and execution across all S8 GEMM variants.
- Swapping A/B for column-major inputs before invoking JIT generator.
- Consolidated JIT kernel initialization and thread decorator invocation
  with transformed inputs instead of branching on column-major flag.
- Fixed B matrix loading offset in S8 GEMM generator.
- Removed S32 > F32 > S32 conversion replicated from classic kernels.
- Update reference kernel to remove F32 cast in final computation.
- Fix mask register count calculation in GEMV generators.
- Remove redundant store path jumps in S8 GEMV generator.

AMD-Internal: [CPUPL-7716]

+++ tests/adaptors/ref/gemm_s8s8s32os32_ref.cc
@@ -123,11 +123,11 @@ aocl_gemm_s8s8s32os32_ref(const char      order,
             sum = dot_product - (128 * b_sum);
 
             if (order == 'R' || order == 'r')
-                C[i * ldc + j] = static_cast<int32_t>(
-                    static_cast<float>((alpha)*sum + (beta)*C[i * ldc + j]));
+                C[i * ldc + j] =
+                    static_cast<int32_t>((alpha)*sum + (beta)*C[i * ldc + j]);
             else
-                C[j * ldc + i] = static_cast<int32_t>(
-                    static_cast<float>((alpha)*sum + (beta)*C[j * ldc + i]));
+                C[j * ldc + i] =
+                    static_cast<int32_t>((alpha)*sum + (beta)*C[j * ldc + i]);
         }
     }
 }
